connect: carve rendezvous.rs (pure move)
The daemon rendezvous leaves main.rs: registration and the claim
lifecycle (codes, co-signed v1/v2 proofs, daemon-signed unclaim), the
daemon poll/answer/error/dry endpoints, dashboard-session bookkeeping,
and the browser offer/ice/close side, plus the claim-code/payload
tests with their daemon_record helper. The tests-only bip39::Language
import rides along; main.rs is under the file-size budget.
